+++ react-fe/src/DocumentPage.js
@@ -1,0 +1,10 @@
+
+function DocumentPage(props) {
+  return (
+    <div>
+      11111
+    </div>
+  );
+}
+
+export default DocumentPage;
